pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "equity-research-agent"
version = "0.1.0"
description = "Agentic equity research drafting workflow using LangGraph, LangChain, Claude, and Azure Blob Storage."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "OpenAI" }
]
dependencies = [
  "langchain>=0.3.0",
  "langchain-anthropic>=0.3.0",
  "langgraph>=0.2.0",
  "azure-storage-blob>=12.24.0",
  "azure-core>=1.32.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "rich>=13.9.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.8.0"
]

[project.scripts]
eq-research-agent = "equity_research_agent.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
